:truck: [#72] Move existing tests to correct module
With the earlier patch I missed we have an existing test module
for input validation.
