+++ src/openforms/submissions/tests/test_submission_step_validate.py
@@ -9,6 +9,7 @@ from rest_framework import status
 from rest_framework.reverse import reverse
 from rest_framework.test import APITestCase
 
+from openforms.config.models import GlobalConfiguration
 from openforms.forms.tests.factories import FormFactory, FormStepFactory
 from openforms.prefill import prefill_variables
 
@@ -320,3 +321,143 @@ class SubmissionStepValidationTests(SubmissionsMixin, APITestCase):
             "invalid_time",
             response.json()["formStep"]["configuration"]["components"][0]["errors"],
         )
+
+    @tag("gh-4143")
+    def test_data_validated(self):
+        """
+        Assert that the shape of data is validated according to the formio definition.
+
+        The validate configuration of each component needs to be applied, but a field
+        being required/optional is irrelevant at this stage (that runs at the end).
+        """
+        submission = SubmissionFactory.create(
+            form__generate_minimal_setup=True,
+            form__formstep__form_definition__configuration={
+                "components": [
+                    {
+                        "type": "textfield",
+                        "key": "textfield",
+                        "label": "Patterned textfield",
+                        "validate": {
+                            "required": True,
+                            "pattern": r"[0-9]{1,3}",
+                        },
+                    },
+                    {
+                        "type": "editgrid",
+                        "key": "repeatingGroup",
+                        "label": "Repeating group",
+                        "components": [
+                            {
+                                "type": "number",
+                                "key": "number",
+                                "label": "Number",
+                                "validate": {
+                                    "required": True,
+                                },
+                            }
+                        ],
+                    },
+                    {
+                        "type": "date",
+                        "key": "dateWithoutValidationInfo",
+                        "label": "Date without validation info",
+                        "validate": {},
+                    },
+                ]
+            },
+        )
+        self._add_submission_to_session(submission)
+        endpoint = reverse(
+            "api:submission-steps-detail",
+            kwargs={
+                "submission_uuid": submission.uuid,
+                "step_uuid": submission.form.formstep_set.get().uuid,
+            },
+        )
+        data = {
+            "textfield": "1a34",  # invalid, because max 3 chars, all numeric
+            "repeatingGroup": [
+                {},  # valid, because required is not enforced
+                {"number": "notanumber"},  # invalid, not a number
+                {"number": None},  # valid: skip required, so `null` is allowed
+            ],
+        }
+
+        response = self.client.put(endpoint, {"data": data})
+
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+        invalid_params = response.json()["invalidParams"]
+        names = {error["name"] for error in invalid_params}
+        expected_names = {"data.textfield", "data.repeatingGroup.1.number"}
+        self.assertEqual(names, expected_names)
+
+    @tag("gh-4143")
+    @patch(
+        "openforms.submissions.api.serializers.GlobalConfiguration.get_solo",
+        return_value=GlobalConfiguration(enable_backend_formio_validation=False),
+    )
+    def test_data_not_validated_with_feature_flag_off(self, m_solo):
+        """
+        Assert that the shape of data is validated according to the formio definition.
+
+        The validate configuration of each component needs to be applied, but a field
+        being required/optional is irrelevant at this stage (that runs at the end).
+        """
+        submission = SubmissionFactory.create(
+            form__generate_minimal_setup=True,
+            form__formstep__form_definition__configuration={
+                "components": [
+                    {
+                        "type": "textfield",
+                        "key": "textfield",
+                        "label": "Patterned textfield",
+                        "validate": {
+                            "required": True,
+                            "pattern": r"[0-9]{1,3}",
+                        },
+                    },
+                    {
+                        "type": "editgrid",
+                        "key": "repeatingGroup",
+                        "label": "Repeating group",
+                        "components": [
+                            {
+                                "type": "number",
+                                "key": "number",
+                                "label": "Number",
+                                "validate": {
+                                    "required": True,
+                                },
+                            }
+                        ],
+                    },
+                    {
+                        "type": "date",
+                        "key": "dateWithoutValidationInfo",
+                        "label": "Date without validation info",
+                        "validate": {},
+                    },
+                ]
+            },
+        )
+        self._add_submission_to_session(submission)
+        endpoint = reverse(
+            "api:submission-steps-detail",
+            kwargs={
+                "submission_uuid": submission.uuid,
+                "step_uuid": submission.form.formstep_set.get().uuid,
+            },
+        )
+        data = {
+            "textfield": "1a34",  # invalid, because max 3 chars, all numeric
+            "repeatingGroup": [
+                {},  # valid, because required is not enforced
+                {"number": "notanumber"},  # invalid, not a number
+                {"number": None},  # valid: skip required, so `null` is allowed
+            ],
+        }
+
+        response = self.client.put(endpoint, {"data": data})
+
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
